Tackle issues of integer width and sign.
* Enable compiler warnings for potentially problematic implicit casts
* Go through code and add explicit casts or avoid conversions
* Switch to using 64-bit integers for adlb IDs and integers on all platforms (i.e. use long long instead of long)
* Also switch data structures to long long to match.
* Add in typedefs for adlb IDs and integers, use the typedefed versions instead of the plain long type
* Use WideInt in tcl/c interface code

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/c-utils/trunk@7269 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ src/exm-string.c
@@ -23,7 +23,7 @@
 bool
 string_copy(char* dest, int d_space, const char* src)
 {
-  int length = strlen(src);
+  size_t length = strlen(src);
   if (d_space <= length)
     return false;
   memcpy(dest, src, length+1);
@@ -35,7 +35,7 @@ string_dup_word(char* src)
 {
   char* q = strchr(src, ' ');
   if (!q) return NULL;
-  int length = q-src;
+  size_t length = (size_t)(q-src);
   char* result = malloc((length+1)*sizeof(char));
   memcpy(result, src, length);
   result[length] = '\0';
@@ -45,7 +45,7 @@ string_dup_word(char* src)
 void
 chomp(char* s)
 {
-  int length = strlen(s) - 1;
+  size_t length = strlen(s) - 1;
   if (length >= 0 && s[length] == '\n')
     s[length] = '\0';
 }
